+++ extensions/omni_schema/tests/fixture/migrations_and_src/migrations/1_migration_no_dependency.sql
@@ -1,0 +1,2 @@
+CREATE TABLE test (data text NOT NULL);
+CREATE TABLE another_test (another_data text NOT NULL);