+++ extensions/omni_schema/tests/fixture/migrations_and_src/migrations/3_another_migration_no_dependency.sql
@@ -1,0 +1,1 @@
+CREATE TABLE third_test (data text NOT NULL check (always_true()));