+++ extensions/omni_schema/tests/fixture/migrations_and_src/src/function_no_dependency.sql
@@ -1,0 +1,7 @@
+create or replace function always_true() returns boolean
+language plpgsql
+as $$
+begin
+    return true;
+end;
+$$;