+++ extensions/omni_schema/tests/fixture/migrations_missing_dependency/migrations/1_migration_depends_function.sql
@@ -1,0 +1,1 @@
+CREATE TABLE test_when_true (data text NOT NULL check (always_true()));