+++ extensions/omni_schema/tests/fixture/non_replaceable_src/migrations/1_migration_no_dependency.sql
@@ -1,0 +1,1 @@
+CREATE TABLE test (data text NOT NULL);